[service-order] Add Integration Test For Order Case

+++ service-order/src/main/java/com/kipaskipas/order/model/Order.java
@@ -4,8 +4,6 @@ import javax.persistence.Id;
 import javax.persistence.JoinColumn;
 import javax.persistence.ManyToOne;
 import javax.persistence.Table;
-
-import java.math.BigDecimal;
 import java.util.Date;
 
 import javax.persistence.Column;
@@ -61,7 +59,7 @@ public class Order {
 
   @NotNull
   @Column(name = "amount")
-  private BigDecimal amount;
+  private Double amount;
 
   @NotNull
   @Column(name = "quantity")
